fix : compilation tree

+++ src/perceptron/perceptron.ipp
@@ -50,7 +50,7 @@ void Perceptron<in, out>::setInput(const std::vector<double>& input)
     }
 
     // Set the inputs
-    for (int index {0}; index < in; ++index)
+    for (size_t index {0}; index < in; ++index)
     {
         *(inputList_[index]) = input[index];
     } 
@@ -75,12 +75,12 @@ double Perceptron<in, out>::getOutput(const int& index)
 template <size_t in, size_t out>
 void Perceptron<in, out>::print() const
 {
-    for (int index1 {0}; index1 < nbLayer_; ++index1)
+    for (size_t index1 {0}; index1 < nbLayer_; ++index1)
     {
-        for (int index2 {0}; index2 < layerList_[index1]; ++index2)
+        for (size_t index2 {0}; index2 < layerList_[index1]; ++index2)
         {
             std::cout << index1 << " - " << index2 << "(" << neuronList_[index1][index2].getInputNbr() << ") : " << neuronList_[index1][index2].getOutputValue() << std::endl;
-            for (int index3 {0}; index3 < neuronList_[index1][index2].getInputNbr() + 1; ++index3)
+            for (size_t index3 {0}; index3 < neuronList_[index1][index2].getInputNbr() + 1; ++index3)
             {
                 std::cout << "W" << std::to_string(index3) << " : " << neuronList_[index1][index2].getWeight(index3) << std::endl;
             }
@@ -98,7 +98,7 @@ void Perceptron<in, out>::learn(const std::vector<double>& input, const std::vec
     for (const auto& layer : layerList_)
     {
         error.emplace_back();
-        for (int index1 {0}; index1 < layer; ++index1)
+        for (size_t index1 {0}; index1 < layer; ++index1)
         {
             error.back().emplace_back(0);
         }
@@ -110,12 +110,12 @@ void Perceptron<in, out>::learn(const std::vector<double>& input, const std::vec
         error.back()[index1] = neuronList_.back()[index1].getDerivativeOutput() * (neuronList_.back()[index1].getOutputValue() - expectedOutput[index1]);
     }
 
-    for (int index1 {nbLayer_ - 2}; index1 >= 0; --index1)
+    for (size_t index1 {nbLayer_ - 2}; index1 >= 0; --index1)
     {
-        for (int index2 {0}; index2 < layerList_[index1]; ++index2)
+        for (size_t index2 {0}; index2 < layerList_[index1]; ++index2)
         {
             double sum = 0;
-            for (int index3 {0}; index3 < layerList_[index1 + 1]; ++index3)
+            for (size_t index3 {0}; index3 < layerList_[index1 + 1]; ++index3)
             {
                 sum += neuronList_[index1 + 1][index3].getWeight(index2) * error[index1 + 1][index3];
             }
@@ -124,11 +124,11 @@ void Perceptron<in, out>::learn(const std::vector<double>& input, const std::vec
     }
 
     // Update all weight
-    for (int index1 {0}; index1 < nbLayer_; ++index1)
+    for (size_t index1 {0}; index1 < nbLayer_; ++index1)
     {
-        for (int index2 {0}; index2 < layerList_[index1]; ++index2)
+        for (size_t index2 {0}; index2 < layerList_[index1]; ++index2)
         {
-            for (int index3 {0}; index3 < neuronList_[index1][index2].getInputNbr(); ++index3)
+            for (size_t index3 {0}; index3 < neuronList_[index1][index2].getInputNbr(); ++index3)
             {
                 neuronList_[index1][index2].setWeight(index3, neuronList_[index1][index2].getWeight(index3) - 
                 lambda * error[index1][index2] * neuronList_[index1][index2].getInputValue(index3));
@@ -143,9 +143,9 @@ template <size_t in, size_t out>
 void Perceptron<in, out>::run()
 {
     // Update all neurons output
-    for (int index1 {0}; index1 < nbLayer_; ++index1)
+    for (size_t index1 {0}; index1 < nbLayer_; ++index1)
     {
-        for (int index2 {0}; index2 < layerList_[index1]; ++index2)
+        for (size_t index2 {0}; index2 < layerList_[index1]; ++index2)
         {
             neuronList_[index1][index2].run();
         }
@@ -156,7 +156,7 @@ template <size_t in, size_t out>
 void Perceptron<in, out>::initialize()
 {
     // Allocation for the entry list
-    for (int index {0}; index < in; ++index)
+    for (size_t index {0}; index < in; ++index)
     {
         inputList_.push_back(std::make_shared<double>(0));
     }
@@ -166,22 +166,22 @@ void Perceptron<in, out>::initialize()
     nbLayer_ = 1;
 
     // Creation of the neurons
-    for (int index {0}; index < out; ++index)
+    for (size_t index {0}; index < out; ++index)
     {
         neuronList_.back().emplace_back(weightedSum, sigmoid, dSigmoid, getRandomNbr(minRdm_, maxRdm_));
     }
 
     // Set the link between entry
-    for (int index1 {0}; index1 < out; ++index1)
+    for (size_t index1 {0}; index1 < out; ++index1)
     {   
-        for (int index2 {0}; index2 < in; ++index2)
+        for (size_t index2 {0}; index2 < in; ++index2)
         {
             neuronList_[0][index1].add(inputList_[index2], getRandomNbr(minRdm_, maxRdm_));
         }
     }
 
     // Allocation for the output list
-    for (int index {0}; index < out; ++index)
+    for (size_t index {0}; index < out; ++index)
     {
         outputList_.push_back(neuronList_[0][index].getOutput());
     }
@@ -204,7 +204,7 @@ void Perceptron<in, out>::addLayer(const size_t& nbNeuron,const int& index)
     // Create the layer
     auto itNeuronList = neuronList_.begin();
     auto itLayerList = layerList_.begin();
-    for (int index1 {0}; index1 < localIndex; ++index1)
+    for (size_t index1 {0}; index1 < localIndex; ++index1)
     {
         ++itNeuronList;
         ++itLayerList;
@@ -214,7 +214,7 @@ void Perceptron<in, out>::addLayer(const size_t& nbNeuron,const int& index)
     ++nbLayer_;
 
     // Create each neurons in the layer
-    for (int index1 {0}; index1 < nbNeuron; ++index1)
+    for (size_t index1 {0}; index1 < nbNeuron; ++index1)
     {
         itNeuronList->emplace_back(weightedSum, sigmoid, dSigmoid, getRandomNbr(minRdm_, maxRdm_));
     }
@@ -223,9 +223,9 @@ void Perceptron<in, out>::addLayer(const size_t& nbNeuron,const int& index)
     if (itLayerList == layerList_.begin())
     {
         // For the first layer
-        for (int index1 {0}; index1 < *itLayerList; ++index1)
+        for (size_t index1 {0}; index1 < *itLayerList; ++index1)
         {   
-            for (int index2 {0}; index2 < in; ++index2)
+            for (size_t index2 {0}; index2 < in; ++index2)
             {
                 itNeuronList->at(index1).add(inputList_[index2], getRandomNbr(minRdm_, maxRdm_));
             }
@@ -233,9 +233,9 @@ void Perceptron<in, out>::addLayer(const size_t& nbNeuron,const int& index)
     }
     else
     {
-        for (int index1 {0}; index1 < *itLayerList; ++index1)
+        for (size_t index1 {0}; index1 < *itLayerList; ++index1)
         {   
-            for (int index2 {0}; index2 < *std::prev(itLayerList, 1); ++index2)
+            for (size_t index2 {0}; index2 < *std::prev(itLayerList, 1); ++index2)
             {
                 itNeuronList->at(index1).add(prev(itNeuronList, 1)->at(index2).getOutput(), getRandomNbr(minRdm_, maxRdm_));
             }
@@ -243,10 +243,10 @@ void Perceptron<in, out>::addLayer(const size_t& nbNeuron,const int& index)
     }
     
     // Set the link with the outputs
-    for (int index1 {0}; index1 < *std::next(itLayerList, 1); ++index1)
+    for (size_t index1 {0}; index1 < *std::next(itLayerList, 1); ++index1)
     {   
         std::next(itNeuronList, 1)->at(index1).clear();
-        for (int index2 {0}; index2 < *itLayerList; ++index2)
+        for (size_t index2 {0}; index2 < *itLayerList; ++index2)
         {
             std::next(itNeuronList, 1)->at(index1).add(itNeuronList->at(index2).getOutput(), getRandomNbr(minRdm_, maxRdm_));
         }
